STRIP_PATH applies before the health check
STRIP_PATH worked great for fixing my AWS application load balancer,
but it didn't help with my health checks.  Prior to this commit,
the health check URL was not subject to STRIP_PATH.  Now it is.

In addition, STRIP_PATH now works as you would expect it to, given
the documentation: it strips the path *prefix*.  Previously, this
would just remove the STRIP_PATH text *anywhere* that it appeared
in the path.